+++ plugins/servicenow-actions/README.md
@@ -18,48 +18,65 @@ Run the following command to install the action package in your Backstage projec
 yarn workspace backend add @janus-idp/backstage-scaffolder-backend-module-servicenow
 ```
 
+### Installing the action on the legacy backend
+
+[Register](https://backstage.io/docs/features/software-templates/writing-custom-actions#registering-custom-actions) the ServiceNow actions by modifying the `packages/backend/src/plugins/scaffolder.ts` file from your project with the following changes:
+
+```ts title="packages/backend/src/plugins/scaffolder.ts"
+/* highlight-add-next-line */
+import { createServiceNowActions } from '@janus-idp/backstage-scaffolder-backend-module-servicenow';
+
+export default async function createPlugin(
+  env: PluginEnvironment,
+): Promise<Router> {
+  // ...
+
+  /* highlight-add-next-line */
+  const actions = [
+    ...builtInActions,
+    ...createServiceNowActions({ config: env.config }),
+  ];
+
+  return await createRouter({
+    actions,
+    logger: env.logger,
+    config: env.config,
+    database: env.database,
+    reader: env.reader,
+    catalogClient,
+    identity: env.identity,
+  });
+}
+```
+
+### Installing the action on the new backend
+
+Add the following to your `packages/backend/src/index.ts` file:
+
+```ts title="packages/backend/src/index.ts"
+const backend = createBackend();
+
+// Add the following line
+backend.add(
+  import('@janus-idp/backstage-scaffolder-backend-module-regex/alpha'),
+);
+
+backend.start();
+```
+
 ## Configuration
 
-1. [Register](https://backstage.io/docs/features/software-templates/writing-custom-actions#registering-custom-actions) the ServiceNow actions by modifying the `packages/backend/src/plugins/scaffolder.ts` file from your project with the following changes:
+Update the `app-config.yaml` file to include the following config:
 
-   ```ts title="packages/backend/src/plugins/scaffolder.ts"
-   /* highlight-add-next-line */
-   import { createServiceNowActions } from '@janus-idp/backstage-scaffolder-backend-module-servicenow';
-
-   export default async function createPlugin(
-     env: PluginEnvironment,
-   ): Promise<Router> {
-     // ...
-
-     /* highlight-add-next-line */
-     const actions = [
-       ...builtInActions,
-       ...createServiceNowActions({ config: env.config }),
-     ];
-
-     return await createRouter({
-       actions,
-       logger: env.logger,
-       config: env.config,
-       database: env.database,
-       reader: env.reader,
-       catalogClient,
-       identity: env.identity,
-     });
-   }
-   ```
-
-2. Update the `app-config.yaml` file to include the following config:
-
-   ```yaml title="app-config.yaml"
-   servicenow:
-     # The base url of the ServiceNow instance.
-     baseUrl: ${SERVICENOW_BASE_URL}
-     # The username to use for authentication.
-     username: ${SERVICENOW_USERNAME}
-     # The password to use for authentication.
-     password: ${SERVICENOW_PASSWORD}
-   ```
+```yaml title="app-config.yaml"
+servicenow:
+  # The base url of the ServiceNow instance.
+  baseUrl: ${SERVICENOW_BASE_URL}
+  # The username to use for authentication.
+  username: ${SERVICENOW_USERNAME}
+  # The password to use for authentication.
+  password: ${SERVICENOW_PASSWORD}
+```
 
 ## Usage
