+++ plugins/servicenow-actions/src/alpha.ts
@@ -1,0 +1,1 @@
+export { scaffolderModuleServicenowActions as default } from './module';
